Refactor to transpiler and move debris to beatmap
