+++ SiraUtil/Objects/Beatmap/Debris/BurstSliderElementNoteLWRegistration.cs
@@ -1,11 +1,11 @@
 ﻿using System;
 
-namespace SiraUtil.Objects.Debris
+namespace SiraUtil.Objects.Beatmap.Debris
 {
     /// <summary>
-    /// Registers a redecorator for the high-quality burst slider element notes (i.e. links in the chain) debris.
+    /// Registers a redecorator for the low-quality burst slider element notes (i.e. links in the chain) debris.
     /// </summary>
-    public sealed class BurstSliderElementNoteHDRegistration : TemplateRedecoratorRegistration<NoteDebris, NoteDebrisPoolInstaller>
+    public sealed class BurstSliderElementNoteLWRegistration : TemplateRedecoratorRegistration<NoteDebris, NoteDebrisPoolInstaller>
     {
         /// <summary>
         /// Creates a new redecorator registration.
@@ -15,6 +15,6 @@ namespace SiraUtil.Objects.Debris
         /// <param name="chain">Whether to chain this redecoration with others. Every redecoration is now aggregated.
         /// The chain will start if the highest priority object has chaining enabled and will stop once a registration
         /// in the aggregate has chaining disabled.</param>
-        public BurstSliderElementNoteHDRegistration(Func<NoteDebris, NoteDebris> redecorateCall, int priority = 0, bool chain = true) : base("_burstSliderElementNoteHDPrefab", redecorateCall, priority, chain) { }
+        public BurstSliderElementNoteLWRegistration(Func<NoteDebris, NoteDebris> redecorateCall, int priority = 0, bool chain = true) : base("_burstSliderElementNoteLWPrefab", redecorateCall, priority, chain) { }
     }
 }
