+++ SiraUtil/Objects/InternalRedecorator.cs
@@ -12,11 +12,11 @@ namespace SiraUtil.Objects
     internal class InternalRedecorator
     {
         private const string NewContextPrefabMethodName = "ByNewContextPrefab";
+        private const string GetContainerMethodName = "get_Container";
         private const string ContainerFieldName = "_container";
         private static readonly MethodInfo _getType = typeof(object).GetMethod(nameof(object.GetType));
         private static readonly MethodInfo _prefabInitializingField = SymbolExtensions.GetMethodInfo(() => PrefabInitializing(null!, null!, null!, null!));
         private static readonly MethodInfo _newPrefabMethod = typeof(FactoryFromBinderBase).GetMethod(nameof(FactoryFromBinderBase.FromComponentInNewPrefab));
-        private static readonly MethodInfo _getContainerMethod = typeof(MonoInstallerBase).GetProperty("Container", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance).GetGetMethod();
 
         [HarmonyPatch(typeof(BeatmapObjectsInstaller), nameof(BeatmapObjectsInstaller.InstallBindings))]
         internal class BeatmapObjects
@@ -42,55 +42,15 @@ namespace SiraUtil.Objects
             }
         }
 
-        // The normal transpiler can't handle this, so we have to use a prefix/postfix.
-        // This was mostly taken from Lapiz, so...
         [HarmonyPatch(typeof(NoteDebrisPoolInstaller), nameof(NoteDebrisPoolInstaller.InstallBindings))]
         internal class NoteDebrisPool
         {
-            private static NoteDebris? _normalNoteDebrisHDPrefabOrig;
-            private static NoteDebris? _normalNoteDebrisLWPrefabOrig;
-            private static NoteDebris? _burstSliderHeadNoteDebrisHDPrefabOrig;
-            private static NoteDebris? _burstSliderHeadNoteDebrisLWPrefabOrig;
-            private static NoteDebris? _burstSliderElementNoteHDPrefabOrig;
-            private static NoteDebris? _burstSliderElementNoteLWPrefabOrig;
-
-            [HarmonyPrefix]
-            protected static void RedecoratePrefix(NoteDebrisPoolInstaller __instance)
+            [HarmonyTranspiler]
+            protected static IEnumerable<CodeInstruction> Redecorate(IEnumerable<CodeInstruction> instructions)
             {
-                DiContainer container = __instance.Container;
-                Type type = __instance.GetType();
-                bool isHD = __instance._noteDebrisHDConditionVariable.value;
-
-                _normalNoteDebrisHDPrefabOrig = __instance._normalNoteDebrisHDPrefab;
-                _normalNoteDebrisLWPrefabOrig = __instance._normalNoteDebrisLWPrefab;
-                _burstSliderHeadNoteDebrisHDPrefabOrig = __instance._burstSliderHeadNoteDebrisHDPrefab;
-                _burstSliderHeadNoteDebrisLWPrefabOrig = __instance._burstSliderHeadNoteDebrisLWPrefab;
-                _burstSliderElementNoteHDPrefabOrig = __instance._burstSliderElementNoteHDPrefab;
-                _burstSliderElementNoteLWPrefabOrig = __instance._burstSliderElementNoteLWPrefab;
-
-                if (isHD)
-                {
-                    __instance._normalNoteDebrisHDPrefab = (NoteDebris)PrefabInitializing(__instance._normalNoteDebrisHDPrefab, container, "_normalNoteDebrisHDPrefab", type);
-                    __instance._burstSliderHeadNoteDebrisHDPrefab = (NoteDebris)PrefabInitializing(__instance._burstSliderHeadNoteDebrisHDPrefab, container, "_burstSliderHeadNoteDebrisHDPrefab", type);
-                    __instance._burstSliderElementNoteHDPrefab = (NoteDebris)PrefabInitializing(__instance._burstSliderElementNoteHDPrefab, container, "_burstSliderElementNoteHDPrefab", type);
-                }
-                else
-                {
-                    __instance._normalNoteDebrisLWPrefab = (NoteDebris)PrefabInitializing(__instance._normalNoteDebrisLWPrefab, container, "_normalNoteDebrisLWPrefab", type);
-                    __instance._burstSliderHeadNoteDebrisLWPrefab = (NoteDebris)PrefabInitializing(__instance._burstSliderHeadNoteDebrisLWPrefab, container, "_burstSliderHeadNoteDebrisLWPrefab", type);
-                    __instance._burstSliderElementNoteLWPrefab = (NoteDebris)PrefabInitializing(__instance._burstSliderElementNoteLWPrefab, container, "_burstSliderElementNoteLWPrefab", type);
-                }
-            }
-
-            [HarmonyPostfix]
-            protected static void RedecoratePostfix(NoteDebrisPoolInstaller __instance)
-            {
-                __instance._normalNoteDebrisHDPrefab = _normalNoteDebrisHDPrefabOrig;
-                __instance._normalNoteDebrisLWPrefab = _normalNoteDebrisLWPrefabOrig;
-                __instance._burstSliderHeadNoteDebrisHDPrefab = _burstSliderHeadNoteDebrisHDPrefabOrig;
-                __instance._burstSliderHeadNoteDebrisLWPrefab = _burstSliderHeadNoteDebrisLWPrefabOrig;
-                __instance._burstSliderElementNoteHDPrefab = _burstSliderElementNoteHDPrefabOrig;
-                __instance._burstSliderElementNoteLWPrefab = _burstSliderElementNoteLWPrefabOrig;
+                List<CodeInstruction> codes = instructions.ToList();
+                InternalRedecorator.Redecorate(ref codes);
+                return codes;
             }
         }
 
@@ -167,8 +127,8 @@ namespace SiraUtil.Objects
                     {
                         for (int c = i; c >= 0; c--)
                         {
-                            // We are in a MonoInstallerBase
-                            if (codes[c].opcode == OpCodes.Call)
+                            // We are in a MonoInstallerBase/ScriptableObjectInstallerBase
+                            if (codes[c].opcode == OpCodes.Call && ((MethodInfo)codes[c].operand).Name == GetContainerMethodName)
                             {
                                 containerOpcode = OpCodes.Callvirt;
                                 containerOperand = codes[c].operand;
